feat: Split WorkflowCycleManager (#20071)
